Rename .reporting → .report
- Rename .report.computations → .report.operator.
- Adjust some usage/references.
