+++ ixmp/report/operator.py
@@ -6,8 +6,9 @@ import pint
 from genno.core.quantity import Quantity
 from genno.util import parse_units
 
-from ixmp.reporting.util import RENAME_DIMS, dims_for_qty, get_reversed_rename_dims
 from ixmp.utils import to_iamc_layout
+
+from .util import RENAME_DIMS, dims_for_qty, get_reversed_rename_dims
 
 log = logging.getLogger(__name__)
